add menu to own file
