+++ cli.py
@@ -1,95 +1,14 @@
 #!/usr/bin/env python
 from spotipy import spotipy
+from spotipy import menu
 import argparse
 import sys
 import curses
-from curses import panel
-
-
-class Menu(object):
-    def __init__(self, items, stdscreen):
-        self.sptfy = spotipy.get_spotipy_class_by_platform()()
-        self.window = stdscreen.subwin(0,0)
-        self.window.keypad(1)
-        self.panel = panel.new_panel(self.window)
-        self.panel.hide()
-        panel.update_panels()
-
-        self.position = 2
-        self.items = items
-        self.items.append(' ')
-        self.items.append('<UP> and <DOWN> to navigate between songs.')
-        self.items.append('<Enter> to select song.')
-        self.items.append('<Esc> for search.')
-        self.items.append('<LEFT> and <RIGHT> for prev/next song.')
-        self.items.append('<SPACEBAR> for play/pause.')
-
-    def navigate(self, n):
-        self.position += n
-
-        if self.position < 2:
-            self.position = 2
-        elif self.position > 16:
-            self.position = 16
-        elif self.position >= len(self.items):
-            self.position = len(self.items) - 1
-
-    def display(self):
-        self.panel.top()
-        self.panel.show()
-        self.window.clear()
-
-        while True:
-            self.window.refresh()
-            curses.doupdate()
-
-            for index, item in enumerate(self.items):
-                if index == self.position:
-                    mode = curses.A_REVERSE
-                else:
-                    mode = curses.A_NORMAL
-
-                self.window.addstr(index, 1, str(item), mode)
-
-            key = self.window.getch()
-
-            if key in [curses.KEY_ENTER, ord('\n')]:
-                if self.position == len(self.items) - 1:
-                    break
-                else:
-                    self.sptfy.listen(int(self.position - 1))
-
-            elif key == curses.KEY_UP:
-                self.navigate(-1)
-
-            elif key == curses.KEY_DOWN:
-                self.navigate(1)
-
-            elif key == curses.KEY_LEFT:
-                self.sptfy.prev()
-
-            elif key == curses.KEY_RIGHT:
-                self.sptfy.next()
-
-            # spacebar
-            elif key == 32:
-                self.sptfy.play_pause()
-
-            # escape
-            elif key == 27:
-                break
-
-        self.window.clear()
-        self.panel.hide()
-        panel.update_panels()
-        curses.doupdate()
 
 
 class App:
     def __init__(self):
         self.sptfy = spotipy.get_spotipy_class_by_platform()()
-
-        self.interaction()
 
         self.run()
 
@@ -101,7 +20,7 @@ class App:
     def menu_items(self, stdscreen):
         curses.curs_set(0)
 
-        main_menu = Menu(self.list, stdscreen)
+        main_menu = menu.Menu(self.list, stdscreen)
         main_menu.display()
